Add Safer Illinois User Guide
added SaferIllinoisUserGuide.pdf
added a navbar item pointing to this file.

I also made a separate User-Guide.html page and uploaded it as an alternative, but it is not linked into the navbar.

I was not able to figure out the purpose of the gtag attribute in the nav link. So I didn't use it.